refactor : 저장매장 바텀시트에 마커 위치 수정 및 파일명 변경
